Refactor common HTML methods into HtmlHelper
- prefix_all_images(html, width)
- giphy_img?(source)
- allowed_image_host?(src)

Offenders:
- app/labor/markdown_parser.rb
- app/models/html_variant.rb
